ART: Fix warnings
Fix Wconstant-conversion and Wundefined-var-template warnings.

Reverts commit df53be273509dd43725870fb20a2c7d71f7fbfd3.

Bug: 28149048
Bug: 29823425
Test: m
Test: m test-art-host

+++ compiler/optimizing/code_generator_x86.cc
@@ -1088,7 +1088,7 @@ void CodeGeneratorX86::GenerateFrameEntry() {
 
   if (GetGraph()->HasShouldDeoptimizeFlag()) {
     // Initialize should_deoptimize flag to 0.
-    __ movl(Address(ESP, -kShouldDeoptimizeFlagSize), Immediate(0));
+    __ movl(Address(ESP, -static_cast<int32_t>(kShouldDeoptimizeFlagSize)), Immediate(0));
   }
 
   int adjust = GetFrameSize() - FrameEntrySpillSize();
